Fix branch_pc if the branch is predicted taken but mispredicted
rv32_fetch sets the pc to branch_pc if the branch was mispredicted.
Previously, branch_pc was always set to the target of the branch -
regardless of whether it was predicted taken or not. If the branch is
predicted taken but is not actually taken then we actually need to set
the pc to the next instruction instead of the target.