Set branch_predicted_taken to 0 if the decode/execute stages are flushed
When a stage is flushed it's supposed to set all its outputs to values
that will make the remaining stages a no-op. However, the combination of
branch_op=NEVER and branch_predicted_taken=1 is not a no-op and causes
the PC to change when it shouldn't.

This commit fixes the bug by ensuring branch_predicted_taken is always 0
when branch_op is set to NEVER during a flush.