Fix rs2_read_out values in the control unit
Some rs2_read_out values were incorrectly set to 2, not 1. 2 in binary
is 10, and this was truncated to 0.

rs2_read_out doesn't need to be set for the read to actually take place
- it's only used for feed forwarding/interlocking, so this only caused
bugs where a sufficiently close prior instruction's rd was set to rs2
(meaning it took me a long time to spot it!)