Resolve branch mispredictions in parallel with instruction fetch
This is to prepare for instruction fetches that take more than a single
cycle. The memory bus doesn't support the address input changing during
a read, so instr_address_out must be set to the output of a register
that doesn't change until the read has finished (this isn't quite true
yet, but will be fixed in a future commit).