Move alu_non_zero calculation to the mem stage
It was originally done at the end of the execute stage because
branch_mispredicted/branch_pc were on the critical path and were
passed directly to the memory bus (which is already a slow part of the
design).

Now that branch misprediction is dealt with in parallel with instruction
fetching, the execute stage is the critical path instead. Moving this
logic to the mem stage allows nextpnr to reach ~35 MHz again.