Move CSR file to the memory access stage
This does mean we need more flip flops (to keep state between the
execute and memory access stage) and a bubble if an instruction
uses a register populated in an immediately prior CSR read.

However, I'm planning to put the logic for generating and returning from
exceptions in the memory access stage (with code for changing the PC and
flushing the pipeline similar to the code for handling mispredicted
branches). As the CSR file mostly stores registers related to exception
handling, I think it makes sense to move it to the memory access stage.

There might also be cases where a CSR write completes if a branch is
mispredicted - I'm not completely sure, but it's safest to move the
parts of instructions with side effects to the memory access stage, by
which time we have committed to retiring the instruction.