Fix verification of the ADDI instruction
The RISC-V formal interface doesn't understand memory access faults, so
it thought the ADDI instruction generated during a memory access fault
caused a fault itself. However, it isn't possible for ADDI instructions
to cause faults, so it fails verification.

This commit fixes the bug by making memory access faults generate an
instruction containing all zeroes, which is still treated as a NOP by
the rest of the pipeline.