Set instr_out to NOP if an instruction fault occurs
It's nice to be able to do this is synthesis for consistency with the
other cases where a bubble is produced. However, it can't be done if the
RISC-V formal interface is enabled, as RVFI thinks the NOP is a valid
ADDI instruction and complains that it didn't expect it to generate an
exception.