Fix logic loop if the bus err signal is used
valid is set to zero if trap is set. However, we need to pass valid into
the load-store unit, whose output is used to determine the trap flag.

This commits adds a valid_before signal representing the validity of the
instruction before execution. We pass this into the load-store unit
instead.