Use valid_before instead of valid in data_hazard()
This reduces the critical path slightly, in theory at the expense of
stalling the decode stage if a trap occurs. However, in reality, if a
trap does occur the instruction in the decode stage will end up being
flushed once the trap reaches the memory access stage, so I don't think
this has a CPI impact in practice.